Made the process of running the integration tests easier.
Now, there are two options for running the integration tests. The first option is as before. The user must install and start OpenDS along with setting the parameters in the testng.xml file.

The second option is recommended and is meant to be a quick way for users to use the integration tests. Here in one command, the OpenDS is installed and started, the testng.xml is automatically generated, and the integration tests are started. Note that this option is quick and efficient, but is inflexible in regards to the parameter settings in the testng.xml file.