Add a new ldifmodify tool that works like ldapmodify except that it alters data in an LDIF file rather than over LDAP.  There are some limitations, including:
- Modify DN operations are not supported
- Operations that would attempt to delete or alter an entry after adding it

Also, this capability has been integrated into the Directory Server startup
process so that it is now possible to have a config/config-changes.ldif that
will automatically be applied to the config/config.ldif file before it gets
processed by the server.  This makes it easy to make configuration changes with
the server offline without directly altering the configuration file, and can be
used to make changes to the configuration needed for running test cases.

OpenDS Issue Numbers:  622, 623