Add a new org.opends.server.interop package where we can place source code that we don't need in the core server but can help other applications interoperate with OpenDS.
Also, add a class org.opends.server.interop.LazyDN which is a
lazily-initialized version of the org.opends.server.types.DN class.  This makes
it possible to have very fast DN processing when you need to create a DN from
a String and the only thing that will be called on that DN is the toString()
method.  This commit includes test cases to ensure that the public methods of
the DN and LazyDN classes stay in sync.

OpenDS Issue Number:  1276