This code introduces creation of a log file in the system's temporary directory for detailed log messages that can be used by any of the quicksetup applications (install, uninstall, upgrade).  Right now logging if fairly meager but we should work on this soon.  This work is being done in anticipation of changes that Brian has suggested in reducing the verbosity of the quicksetup applications.  The idea is to put the details in the file and omit the messages in the progress panel details area (for instance messages about moving files around).
The file is enabled using static methods in QuickSetupLog.  All that is necessary for applications that want to write to a log file is to call initLogFile() passing in a File object.  This will need to be done differently for the CLI and GUI applications given the differences in how they are launched (see the changes to the Upgrader and UpgradeLauncher for details on this).

If the log file has been initialized, CLI application automatically print a message at the end of execution notifying the user of the log file.  GUI applications need to manually call notifyListenersOfLog() to get the message to the details area of the progress panel.