single valued attribute conflict resolution : issue 609
The resolution procedure for single valued attribute needs to be slightly different
from the mult-valued attribute procedure :
- less historical information can be kept
- the procedure must take into account the fact that only one value is allowed
 at a given time.

This Change splits the AttrInfo class into 2 classes :
AttrInfoSingle and AttrInfoMultiple that both extends AttributeInfo.
The Historical class if also refactorized to become more generic, some code
was staying there but was indeed specific to multi-valued attribute.

This change also add a number of unit tests for single valued attribute, and enable
an old test from HistoricalTest.java that was previously disabled because
conflict resolution for single valued attribute was not yet implemented.