Fix the replication WindowProbeTest and enable back this test.
The problem was that because of some tests running before this test
they were changes in the replication server when this test was run.
The test now open a first session to read the serverState of the 
replication server then open a second session using this serverState
as its own.