Fix for issue 1321 : Synchronization changes aren't logged by default
This adds a  new configuration property for the accessLogPublishers :  ds-cfg-suppress-synchronization-operations.
It can be used to configure the behaviour of some log publisher with regards to the synchronization operations.

When set to false (which is the default value) the access log publisher will log the synchronization operation.
When set to true, the access log publisher will not log the synchronization operation.
The property ds-cfg-suppress-internal-operations that control the logging of the internal operations
now does not have any effect on the synchronization operation (so that one can configure the
server to log all internal operations except the synchronization ones)

Both the access log and the audit log implement this property.

I've also added the string : "type=synchronization" at the end of the line corresponding
to the operation request in the access log so that it is possible to
differentiate the synchronzation operations from the other operations.

example :
[10/Jul/2007:15:29:34 +0200] ADD conn=-1 op=1314 msgID=1315 dn="cn=Natascha Pfifferling463, ou=Human Resources, dc=example,dc=com" type=synchronization
[10/Jul/2007:15:29:34 +0200] ADD conn=-1 op=1314 msgID=1315 result="Success" etime=98

I could not find any unit test for the access and audit log and decided it was too
complex to create new ones compared to the simplicity of this change...