Fix issues 1558 and 1919.
Modify the LDAPProfile API so that it is possible to decorate the core LDAPProfile. When a LDAPProfile query is received each decorator is queried in sequence until one can return a response. If none of the decorators can respond, then the core LDAP profile is queried.

This change makes it easier to create test configuration definitions in the unit tests. A unit test can decorate the core LDAPProfile with additional profile information for its test configuration definitions.

In addition, we now have test-parent and test-child components which are for use in unit tests. They are generated from XML definitions in a similar manner to other components (this must be done manually though at the moment).

Whilst updating the admin LDAP client test suite I encountered a bug in inherited default value resolution. This has also been fixed.