Partial 95% fix for issue 1831 - dsconfig interactive mode.
This change implements the remaining functionality for the dsconfig interactive mode. It is now possible to interactively create and modify components. This change adds support for interactively querying the user for property values. It is implemented using a text-based menu driven model, whereby users can select which properties they want to modify and how they want to modify them (e.g. reset, add a value, remove a value, etc).

This change is not a complete fix for issue 1831, since there are some minor usability enhancements that can be made. These include:

* general consistency of menus (e.g. being able to cancel and go back, get help on a property, etc)

* when displaying the menu option for resetting a property to its defaults it should display what those are exactly

* when incrementally modifying a property, it would be nicer if it displayed the current state

The aim of this change is principally to get people to use the interactive mode and to identify any major usability design issues.