Fix for issue 2059.
Mark the panels as non opaque to avoid having the white box in certain windows environments.

Instead of simulating a click on the OK button when the user accepts the certificate, directly call the method invoked when the button is clicked.  This avoids having issues with focus, that actually make the button click to be ignored.

Use hexadecimal representation of the certificate instead of the address of the byte array.