This set of changes add ID to identify network groups, workflows and workflow elements. These identifiers pave the way for the network group and workflow configuration.
A workflow that is registered with a network group must have an identifier that is not already registered with the network group. To deregister a workflow one has to provide either a workflow base DN or a workflow identifier. The former method is used when a backend is deregistered with the server while the later may be used when applying workflow configuration changes.

The unit-tests are updated accordingly.