Add some log messages in the initialization code.
Fix a bug in that made the progress bar to go back.

Use LinkedHashSets instead of HashSets in SuffixesToReplicateOptions and return a copy of these Sets.  This is done to be sure that the ADS is the last suffix to be initialized.

Use a customized message when we are initializing the ADS suffix instead of displaying the suffix (the user might be surprised by the fact that a base DN (s)he never chose is being initialized).