Add a new LDIF connection handler, which can be used to watch for new LDIF files to be created in a directory, at which point the changes contained in those LDIF files will be processed by the server using internal operations. This can potentially be used to make configuration changes or invoke tasks if the server is unresponsive to external clients (e.g., all worker threads are busy).
Note that in the default configuration, the LDIF connection handler is enabled
but the target directory doesn't exist.  In this configuration, it will simply
not do anything until the target directory is created and one or more LDIF
files are placed in it.  This is a safeguard against an administrator
inadvertently placing a file in that directory without understanding the
potential consequences of that action.  It also means that we can use it to
perform debugging operations (e.g., "dump all") without having to restart the
server if it becomes unresponsive.

OpenDS Issue Number:  2216