The following changes in the replication server provide an implementation of the backend API for the replication server changes database. In a first step, the implemented features are backup/restore. The coming ones will be export/search to have an LDAP access of the content of the replication server DB.
I also cleaned/shared some code on replication unit tests.
Also fixed a potential bug in the replication plugin on the total update by replacing any object dependency from the ReplicationDomain to the associated 
backend because the configuration changes on the backend object may replace this object by another instance, so keeping a reference on the backend instance is 
buggy. The instance must be retrieved when needed.