Modify the aggregation definition property stylesheet to generate two server-side getter methods for each aggregation property. The first is a "default" getter of the form:
    String getXXX()
    SortedSet<String> getXXX()

Where XXX is the name of the property. This getter follows the same pattern as all other property getters. In this case it returns a String (or set of Strings) containing the common name(s) of the referenced component(s).

The second server-side getter is of the form:

    DN getXXXDN()
    SortedSet<DN> getXXXDNs()

Where again XXX is the name of the property. This getter returns the distinguished name(s) associated with the referenced component(s). It is useful on the server-side where component implementations need to retrieve referenced implementations from a DirectoryServer table based on their DN.