Fix for issue 2214 (Provide a manner to accept permanently certificates).
Change the labels of the Certificate Dialog and add a new button to be able to update the certificate permanently.

The modifications only apply to graphical interface since the command-line utilities already allow the user to accept permanently the certificates.

These modifications include also a fix in ApplicationTrustManager which did not parse properly the Subject DN of the certificate and reported systematically a host name mismatch.