There was a flag day when the new security framework was added in build 1.0.0 build 6.  The following changes address the issue.  The main thing that has been updated are the changes at the end of VersionCompatibilityIssue.  The problem is that the code relied only on the fact that a flag day was known (we register the flag days).  However it can happen (as it is the case) that a flag day is registered in the code way after the changes occurred.  So when we registered a flag day, systematically we could not upgrade from previous builds.
For instance: a flag day corresponding to build 6 is registered now (candidate build 11). Without these modifications this would prevent to upgrade from build 7 or build 8 to the current build.  That is why the new code actually checks the revision corresponding to the flag day and the revision of the new build.  If the new build is newer than the revision of the flag day then we consider that there is no problem and we continue with the upgrade.