Fix for issue 2902 (stop-ds -R restarts the server with client tool JVM)
When the java properties are set to not overwrite the environment variables (which is the default behavior) when we call start-ds from the stop-ds scripts, OPENDS_JAVA_ARGS is already set with the arguments for the stop (client arguments).  The fix consists basically on resetting the environment variables to those we had before calling stop-ds before calling start-ds.