Fix for 2968 : dependent ADD operation can incorrectly generate conflicting entries
With the following sequence of operation :
Add first entry with DN X
delete first entry with DN X
Add another entry with DN X

The second ADD is sometimes not computed as dependent and a false
DN conflict is sometime created.

This happens when the delete operation complete after we tried to process the second
add but before we check for dependencies in the operation queue.

To avoid this problem all operations must be attempted twice before going into
the conflict resolution mechanism, unfortunately in this case, we were not checking if
this is the first try he operation is attempted and therefore were going directory into
conflict resolution.