Fix for issue 2998 (Command ./bin/create-rc-script should not hard-code the OPENDS_JAVA_HOME in the script)
If the user specifies in the java.properties not to overwrite the environment the java properties specified in java.properties will take precedence over the variables set in the script.  In general users using java.properties should use the overwrite mode (not doing it does not make much sense).

I have updated the code to set the OPENDS_JAVA_HOME in the rc script only if the user explicitly sets the java home in the create rc command line or if it is already set in the environment (and not systematically as it is today).

A bug with the name used for the environment variable of the java properties has also been fixed.