Fix for issue 3049 (./uninstall --cli -a --no-prompt exit status 0 (success) despite errors reported)
Throw an UserDataException when the user must provide authentication and we are in non-interactive mode without the forceOnError flag.  This way we display the error message describing the problem, the alternatives in terms of arguments that can be used and the whole usage of the command line with a longer description of the arguments we propose to use.