Fix for issue 2197 (dsconfig: interactive mode should display the effective dsconfig command)
After discussing with different members the following has been committed:

* The current behavior is kept (no additional output by default).

* Two new options have been added:
--displayEquivalentCommand
    Display the equivalent non-interactive argument in the standard output when
    this command is run in interactive mode
--equivalentCommandFilePath {path}
    The full path to the file where the equivalent non-interactive
    commands will be written when this command is run in interactive mode


So the user will have to explicitly ask to display the equivalent command-line (or to dump it to a file).  This approach privileges the users that configure the server using interactive dsconfig instead of those whose ultimate goal is to create scripts to configure the server (and to whom this feature is targetted).

In terms of implementation basically what is done is to create a new class called CommandBuilder when we store the equivalent arguments to what the user provides interactively.  When the user applies a change (or asks to display a list of objects) the equivalent command-line is displayed by using the contents stored in the CommandBuilder object.