Fix for issue #3958: Upgrade should support products built on top of OpenDS
OpenDS can be customized: quicksetup splashscreen / product name ... But upgrade CLI fails as soon as current version and target version are the same. This change allows upgrade to a same version as soon as the product names are different so allows upgrading from an OpenDS to a customized OpenDS.

This also fixes issue #3959 (Add a way to have the version of the instance) for Unix platform. For that, we add a file in the instance files (config/buildinfo) that contains the version number of the instance. This file were already created in SVR4 and pkg(5) deliveries context. This change creates the file in ZIP delivery context.
For Windows platform, some changes are required in the .bat files to solve issue #3959 (Add a way to have the version of the instance). These changes will be handled by another commit.