Fix for issue #3960: add a way to have the version of the instance (Windows part)
This fixes issue #3960 (add a way to have the version of the instance) for Windows platforms by  managing a buildinfo file at instance level and modifying CLIs to protect use of them while binaries and instance are not in sync.
Note that this behaviour is already implemented on Unix platforms since 1.2 for OpenSolaris delivery and rev 5301 for ZIP delivery.