Partial Fix for issue #3938 : dsconfig is really too slow in non-interactive mode, almost unusable on Niagara systems
This is a partial fix for this dsconfig performance issue:
Linked to fix for issue #2017 (tools usage should disclose default value for options),
we systematically parsed the configuration file.
This fix allows to only parse the config file when necessary (i.e. for help command, or if -port/--trustAll not specified in the command line)

On a machine with a CPU almost equivalent of the Niagara ones,
it allows to save ~20% of the dsconfig time (~9s instead of ~11s).