Fix for 2.0 branch : issue 3995 : hitting queue-size-bytes max cause replication to stop
While running replication tests with 8 replicas the replication
can hang because of some incoherency inside the Replication Server
code that cause the size of the internal queue to become wrong.

After several tests it looks like that this happens because the read
access to the internal Map that is used to implement the queue are
not protected by the lock that already protect the add and delete access.

The changes provided here therefore add locking to all the access to
the map and also add some tracing that would allow to detect the problem
should it happen again.