Partial fix for 3250 : dsreplication init failed routing table is empty
The replication init can fail because the heartbeat timer is set to a default
value of 1 sec and when the JVM is not tuned to achieve low garbage collection
time the GC can last several seconds during which the whole process is stopped
and this can therefore trigger the heartbeat timer on remote servers.

The real fix would be to make the init protocol resilient to reconnection but
this is too late for 2.0.
For the 2.0 the default value of the eartbeat timer is therefore set to 10 sec

The downside of using a higher value for the heartbeat timer is that in case
of a silent hang of a Replication Server a Directory Server can be stopped from
processing changes during the same amount of time. Hopefully this should
be very rare.

Users needing to use lower values of heartbeat timer can still achieve it by
tuning the garbage collector in order to achieve low pause time and changing
the heartbeat value.

This needs to be documented.