Fix for issue 3996 : Changes (DEL) Lost inside a replication topology
The changes are lost because the Directory Server that process the changes is
connected to a Directory Server that is periodically shutdown.

In such cases some Operations might not have yet been transmitted to other
Replication Servers and can therefore be lost.

This is a known issue (see 1752) but it appear that it can appear in conditions
that are more frequent that what I was thinking.

These diffs attempt to fix the problem by.
- giving the preference to the local Replication Server if there is one available and it
  is up to date.
- keep a list of DELETE operation in memory and retransmit them when reconnecting
  to a Replication Server that is missing some changes

This is a partial fix but QA has successfully tested that it fixes his problem.

I will keep 1752 open so that we build a better fix in the trunk