fix for issue 4028 : error 53 returned to application during reconnection
When replication is configured and when the connection between the Directory Server
and the Replication Server fails the Directory Server is reconnecting to a Replication Server.
However currently the error 53 (DSA is unwilling to perform) is returned to application
during the time of the reconnection.
This is wrong because at this time the DS is not on error.

With this change the update threads now waits for the new connection to
be established (which is fast) before being processed.