Fix for issue 3912 (Default automatic Backup should be offered by the control panel)
The fix consists on adding three new elements to the control panel.

1. Allow the user to specify a future date when the backup will be launched or a periodic schedule to launch the backup.

2. Allow the user to specify a future date when an export to LDIF will be launched or a periodic schedule to launch the export.

3. Allow the user to see the scheduled and ongoing tasks.  This is a graphical equivalent of the manage-task command.  In this first version, the GUI displays the tasks and allows to cancel them.  It allows to display all the common task attributes in a table and also the specific attributes for a task in a separate pane.

With this proposal including the possibility of allowing scheduling other type of tasks (like import or restore) is just a matter of adding half a dozen of lines of code (since most of the code is in StatusGenericPanel) so if we find that allowing scheduling tasks for other tasks (apart from backup and export) makes sense, it is pretty straightforward to add this functionality.