Fix for issue 4080 (Using Quicksetup jnlp link: OpenDS splashscreen is displayed during a long while)
The problem comes because quicksetup.jar contains the utility properties.  The code tries to load the correct bundle for the preferred language (for ES), since it is not in quicksetup.jar (no messages_en is included), waits till all the jars are downloaded.  This makes the splashscreen to stay till all the jars are downloaded.

In fact this can be considered a duplicate of  2112.  The fix consists on marking utility.properties the same as quicksetup.properties: when the server is launched using java web start it will use a particular class loader.