Fix for issue 4148 (setup fails : The import LDIF tool returned error code 1) Since the import-ldif utility does not support any more (temporarily) the append option the setup (and the control-panel) were broken.
Temporarily the following modifications have been made in the control panel:
If the user decides to create data in the new base DN and there was already data in the backend, a confirmation dialog will be displayed informing that all the data will be overwritten.
The appending options in the Import LDIF panel have been hidden.

Concerning the setup, the modifications made are permanent.  What basically has been done is:
Use a single template LDIF file when the user decides to import automatically generated data.  Before this, one file per base DN was generated, which required the option 'append' to be used.  With a single file, no append is required.  This is a clearer solution, since I think that appending data affects to performance.
I have found some memory issues with the default ergonomics of the new import (see issue 4151).  Since some ergonomics issues may occur, the import is now launched in a separate process.  If we manage to improve the code that determines the default java arguments to be used for the import, this will help avoiding problems with the import.