Fix for issue 4295 (Issues with dsreplication interactive mode)
The biggest issue was that the trust manager was reset when connecting to the second server, this caused the establishing of the connections to fail.

There was also a problem with the way the code found out whether a NamingException was caused by a certificate problem or not.

Finally the code handles the particular case where the user provides global administrator credentials but this is not defined on the second server.  If this is the case, the code now prompts only for the credentials (and not for the whole second server connection parameters).