Mention the need for enough file descriptors in the release notes chapter on overall prerequisites, rather than the CLI install procedure. The default settings on recent CentOS and Ubuntu systems seem to be 1024/user, which is not enough for OpenDJ in production.
It makes sense to call attention to this as an overall requirement because the admin needs root or sudoer access to edit /etc/security/limits.conf, and the rest of the CLI install can be done by a normal user. Also, if for some reason production installs are done with the GUI, then the information could be missed if only in the CLI install procedure.